+++ R/rpanet_general.R
@@ -22,16 +22,16 @@ NULL
 #' Generate a PA network with non-linear preference functions
 #'
 #' @param nstep Number of steps when generating a network.
-#' @param seednetwork A list represents the seed network. If \code{NULL},
-#'   \code{seednetwork} will have one edge from node 1 to node 2 with weight 1.
+#' @param initial.network A list represents the seed network. If \code{NULL},
+#'   \code{initial.network} will have one edge from node 1 to node 2 with weight 1.
 #'   It consists of the following components: a two column matrix
 #'   \code{edgelist} represents the edges; a vector \code{edgeweight} represents
 #'   the weight of edges; a integer vector \code{nodegroup} represents the group
 #'   of nodes. \code{nodegroup} is defined for directed networks, if
 #'   \code{NULL}, all nodes from the seed graph are considered from group 1.
 #' @param control A list of parameters that controls the PA generation process.
-#'   The default value is \code{rpacontrol.scenario() + rpacontrol.edgeweight() +
-#'   rpacontrol.newedge() + rpacontrol.preference() + rpacontrol.reciprocal()}. By
+#'   The default value is \code{rpa_control_scenario() + rpa_control_edgeweight() +
+#'   rpa_control_newedge() + rpa_control_preference() + rpa_control_reciprocal()}. By
 #'   default, in each step, a new edge of weight 1 is added from a new node
 #'   \code{A} to an existing node \code{B} (\code{alpha} scenario), where
 #'   $\code{B} is chosen with probability proportional to its in-strength + 1.
@@ -40,8 +40,8 @@ NULL
 #' @param m Integer vector, number of new edges in each step.
 #' @param sum_m Integer, summation of \code{m}.
 #' @param w Vector, weight of new edges.
-#' @param nnode Integer, number of nodes in \code{seednetwork}.
-#' @param nedge Integer, number of edges in \code{seednetwork}.
+#' @param nnode Integer, number of nodes in \code{initial.network}.
+#' @param nedge Integer, number of edges in \code{initial.network}.
 #' @param method Which method to use when generating PA networks: "binary" or
 #'   "linear".
 #' @param sample.recip Whether reciprocal edges will be added.
@@ -53,25 +53,25 @@ NULL
 #'   control list
 #'   \code{control}, node group \code{nodegroup} (if applicable) and edge
 #'   scenario \code{scenario} (1~alpha, 2~beta, 3~gamma, 4~xi, 5~rho,
-#'   6~reciprocal). The scenario of edges from \code{seednetwork} are denoted as
+#'   6~reciprocal). The scenario of edges from \code{initial.network} are denoted as
 #'   0.
 #'   
 #' @keywords internal
 #' 
-rpanet_general <- function(nstep, seednetwork, control, directed,
+rpanet_general <- function(nstep, initial.network, control, directed,
                            m, sum_m, w,
                            nnode, nedge, method, sample.recip) {  
-  edgeweight <- c(seednetwork$edgeweight, w)
+  edgeweight <- c(initial.network$edgeweight, w)
   node_vec_length <- (sum_m + nedge) * 2
   node_vec1 <- integer(node_vec_length)
   node_vec2 <- integer(node_vec_length)
   scenario <- integer(node_vec_length)
-  node_vec1[1:nedge] <- seednetwork$edgelist[, 1] - 1
-  node_vec2[1:nedge] <- seednetwork$edgelist[, 2] - 1
+  node_vec1[1:nedge] <- initial.network$edgelist[, 1] - 1
+  node_vec2[1:nedge] <- initial.network$edgelist[, 2] - 1
   scenario[1:nedge] <- 0
-  seed_strength <- node_strength_cpp(seednetwork$edgelist[, 1], 
-                                     seednetwork$edgelist[, 2], 
-                                     seednetwork$edgeweight,
+  seed_strength <- node_strength_cpp(initial.network$edgelist[, 1], 
+                                     initial.network$edgelist[, 2], 
+                                     initial.network$edgeweight,
                                      nnode, weighted = TRUE)
   control$preference$ftype.temp <- ifelse(control$preference$ftype == "default", 
                                           yes = 1, no = 2)
@@ -87,8 +87,8 @@ rpanet_general <- function(nstep, seednetwork, control, directed,
       control$reciprocal$recip.prob <- matrix(0)
     }
     nodegroup <- integer(node_vec_length)
-    nodegroup[1:nnode] <- seednetwork$nodegroup - 1
-    # nodegroup <- c(seednetwork$nodegroup - 1, integer(node_vec_length))
+    nodegroup[1:nnode] <- initial.network$nodegroup - 1
+    # nodegroup <- c(initial.network$nodegroup - 1, integer(node_vec_length))
     if (method == "binary") {
       ret_c <- rpanet_binary_directed(nstep,
                                       m,
@@ -192,7 +192,7 @@ rpanet_general <- function(nstep, seednetwork, control, directed,
               "scenario" = ret_c$scenario[1:nedge], 
               "newedge" = ret_c$m,
               "control" = control,
-              "seednetwork" = seednetwork[c("edgelist", "edgeweight", "nodegroup")], 
+              "initial.network" = initial.network[c("edgelist", "edgeweight", "nodegroup")], 
               "directed" = directed)
   colnames(ret$edgelist) <- NULL
   if (directed) {
@@ -211,7 +211,7 @@ rpanet_general <- function(nstep, seednetwork, control, directed,
   else {
     ret$control$reciprocal$group.prob <- NULL
     ret$control$reciprocal$recip.prob <- NULL
-    ret$seednetwork$nodegroup <- NULL
+    ret$initial.network$nodegroup <- NULL
   }
   return(ret)
 }